Add test for NaiveDateTime in tests module
This commit introduces a new test function, `test_date_time`, to validate the `NaiveDateTime` functionality of the `chrono` crate. Additionally, the `NaiveDateTime` import has been added to the list of imports from the `chrono` library to support this test. The test verifies year, month, day, hour, minute, second, and nanosecond components of the `NaiveDateTime`.